+++ maru/executionlayer/src/main/kotlin/maru/executionlayer/client/CancunWeb3JJsonRpcExecutionLayerEngineApiClient.kt
@@ -10,9 +10,9 @@ package maru.executionlayer.client
 
 import maru.consensus.ElFork
 import maru.core.ExecutionPayload
-import maru.extensions.captureTimeSafeFuture
-import maru.mappers.Mappers.toDomainExecutionPayload
-import maru.mappers.Mappers.toExecutionPayloadV3
+import maru.executionlayer.mappers.Mappers.toDomainExecutionPayload
+import maru.executionlayer.mappers.Mappers.toExecutionPayloadV3
+import net.consensys.linea.async.toSafeFuture
 import net.consensys.linea.metrics.MetricsFacade
 import org.apache.tuweni.bytes.Bytes32
 import tech.pegasys.teku.ethereum.executionclient.schema.ForkChoiceStateV1
@@ -34,7 +34,7 @@ class CancunWeb3JJsonRpcExecutionLayerEngineApiClient(
   override fun getFork(): ElFork = ElFork.Cancun
 
   override fun getPayload(payloadId: Bytes8): SafeFuture<Response<ExecutionPayload>> =
-    createRequestTimer<ExecutionPayload>(method = "getPayload").captureTimeSafeFuture(
+    createRequestTimer<ExecutionPayload>(method = "getPayload").captureTime(
       web3jEngineClient.getPayloadV3(payloadId).thenApply {
         when {
           it.payload != null ->
@@ -47,10 +47,10 @@ class CancunWeb3JJsonRpcExecutionLayerEngineApiClient(
             throw IllegalStateException("Failed to get payload!")
         }
       },
-    )
+    ).toSafeFuture()
 
   override fun newPayload(executionPayload: ExecutionPayload): SafeFuture<Response<PayloadStatusV1>> =
-    createRequestTimer<PayloadStatusV1>(method = "newPayload").captureTimeSafeFuture(
+    createRequestTimer<PayloadStatusV1>(method = "newPayload").captureTime(
       web3jEngineClient
         .newPayloadV3(
           /* executionPayload = */
@@ -66,7 +66,7 @@ class CancunWeb3JJsonRpcExecutionLayerEngineApiClient(
             Response.fromErrorMessage(it.errorMessage)
           }
         },
-    )
+    ).toSafeFuture()
 
   override fun forkChoiceUpdate(
     forkChoiceState: ForkChoiceStateV1,
@@ -74,9 +74,9 @@ class CancunWeb3JJsonRpcExecutionLayerEngineApiClient(
   ): SafeFuture<Response<ForkChoiceUpdatedResult>> =
     createRequestTimer<ForkChoiceUpdatedResult>(
       method = "forkChoiceUpdate",
-    ).captureTimeSafeFuture(
+    ).captureTime(
       web3jEngineClient.forkChoiceUpdatedV3(forkChoiceState, Optional.ofNullable(payloadAttributes?.toV3())),
-    )
+    ).toSafeFuture()
 
   private fun PayloadAttributesV1.toV3(): PayloadAttributesV3 =
     PayloadAttributesV3(